refactor(tests): refactorizar tests de módulo Orders - dividir archivos >300 líneas
- Dividir order-processing-saga.service.spec.ts (867 líneas) en 5 archivos especializados
- Dividir orders.controller.spec.ts (413 líneas) en 3 archivos especializados
- Dividir orders.service.spec.ts (329 líneas) en 2 archivos especializados
- Extraer helpers comunes a order-saga.test-helpers.ts (260 líneas)
- Extraer helpers comunes a orders.test-helpers.ts (150 líneas)
- Implementar test.each() para validaciones de status en controller
- Mantener coverage de 73.73% (Orders: 86-97%)

Archivos creados:
- order-processing-saga.service.core.spec.ts: definición de servicio, startOrderProcessing, success path, circuit breaker (180 líneas)
- order-processing-saga.service.compensations.spec.ts: REFUND_PAYMENT, NOTIFY_FAILURE, compensation failures (200 líneas)
- order-processing-saga.service.failures.spec.ts: stock failures, payment failures, saga not found, non-retryable errors (180 líneas)
- order-processing-saga.service.retries.spec.ts: exponential backoff, max retries, notification failures (150 líneas)
- order-processing-saga.service.edge-cases.spec.ts: multiple product items, circuit breaker states (120 líneas)
- orders.controller.core.spec.ts: definición, createOrder, getUserOrders, getOrderById, getOrderStatus (140 líneas)
- orders.controller.validations.spec.ts: single/multiple items, user variations, status variations con test.each() (100 líneas)
- orders.controller.errors.spec.ts: error propagation para todos los endpoints (100 líneas)
- orders.service.core.spec.ts: definición, createOrder success, idempotency, findUserOrders, findOrderById, getOrderStatus (250 líneas)
- orders.service.errors.spec.ts: product not found, transaction rollback, order not found, unauthorized access (170 líneas)
- helpers/order-saga.test-helpers.ts: factories y assertion helpers para saga tests (260 líneas)
- helpers/orders.test-helpers.ts: factories y assertion helpers para orders tests (150 líneas)

Tests: 50 passing, Coverage: 73.73% general (Orders: 86.62-97.5%)
